% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/age_dist.R
\name{get_pop_by_age}
\alias{get_pop_by_age}
\title{Get Census Counts for Any Country}
\usage{
get_pop_by_age(countries, years, ages = 15, agg = "aggregate",
  rvalue = -1L)
}
\arguments{
\item{countries}{Uses gsub on URL.  Can be vector of names (eg, c("US", "UK") and will group into region if A=aggregate.}

\item{years}{Uses gsub on URL.  Can be vector (eg, c(2015,2016,2017,2018)) and will stack years}

\item{ages}{is set at 10 for 5 year age groups and 15 for 1 year age groups (default)}

\item{agg}{This defaults to "aggregate" at the moment (other option is "separate").}

\item{rvalue}{R = 1 or -1 (Need to check this one for what it does.  Set at -1 for now.)}
}
\value{
Results in a dataframe with one row per age. Note that even with 1 year ages, 85 is 85-89, 90 is 90-94, 95 is 95-99, and 100 is 100 and older.
}
\description{
Gets population counts by age for any country using a 2-letter country code, as provided
by the US Census international data.
}
\details{
Derived from original code at https://raw.githubusercontent.com/walkerke/teaching-with-datavis/master/pyramids/rcharts_pyramids.R
}
\examples{
\dontrun{
pop1 <- get_pop_by_age("US", 2015)
saveRDS(pop1, "./data/analysis/us_2015_1yr.rds")
}
}

