% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_files}
\alias{load_files}
\title{Load SEER files}
\usage{
load_files(fname = ".", vec_of_files)
}
\arguments{
\item{fname}{Name of the SEER file to retrieve.  Uses grepl on a list of the SEER filenames so the string can be written flexibly.}

\item{vec_of_files}{Vector of filenames to search through.  Usually provided by listing filenames ending in .txt in SEER data directory.}
}
\value{
This function returns a dataframe of the files across SEER years.
}
\description{
This function gets files by name and binds them into a data.frame.
Selects the datafile names to select the files to merge using a logical vector to select files.  Data files are "BREAST", "COLRECT", "DIGOTHR", "FEMGEN", "LYMYLEUK", "MALEGEN", "RESPIR", "URINARY", and "OTHER" (names do NOT have to be in caps).
}
\examples{
\dontrun{
# get AML data and use data.table to subset and select.  Need to run setup_incidence() first.
s <- setup_incidence()
seer <- load_files("LYMYLEUK", s$datafiles)
aml <-
    setDT(seer) \%>\%
    setkey(., siterwho) \%>\%
    .[J(35021)] \%>\%
    .[, type := "aml"]
setkey(aml, casenum)
aml <- unique(aml) # removes duplicate aml records (only taking first of each)
}
}

