% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_sas_specs}
\alias{read_sas_specs}
\title{Read SAS input statement}
\usage{
read_sas_specs(textdoc, ...)
}
\arguments{
\item{textdoc}{The text document to be read in.  Must be of a very specific format.}

\item{...}{To pass parameters to the readLines() function.}
}
\value{
Returns specifications from SAS input statements that can be used by R
}
\description{
This function reads the SEER SAS input statements file to create fwf input specs.
}
