% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{setup_incidence}
\alias{setup_incidence}
\title{Setup to Load Fixed-width Incidence Files}
\usage{
setup_incidence(inc_dir = "./data/raw/", yr = 2013)
}
\arguments{
\item{inc_dir}{Directory on user's system that contains the incidence data.}

\item{yr}{Year of release.  Defaults to 2013 (2016 release).  Note that incidence function is for 2013 only.}
}
\value{
xxx
}
\description{
Reads in file specs and sets up information to load desired incidence files.  These files come in groups, and specific tumors are within the groups.
}
\examples{
\dontrun{
# simple script read in files with types using above inputs
# remove col_types to have it choose types automatically
s <- setup_incidence()
seer <- load_files("LYMYLEUK", s)
}
}

