% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{setup_incidence}
\alias{setup_incidence}
\title{Setup to Load Fixed-width Incidence Files
Reads in file specs and sets up information to load desired incidence files.}
\usage{
setup_incidence(inc_dir = "./data/raw/")
}
\arguments{
\item{inc_dir}{Directory on user's system that contains the incidence data.  Defaults to current year (2015 release of 2012 data)}
}
\value{
xxx
}
\description{
Setup to Load Fixed-width Incidence Files
Reads in file specs and sets up information to load desired incidence files.
}
\examples{
\dontrun{
# simple script read in files with types using above inputs
# remove col_types to have it choose types automatically
s <- setup_incidence()
seer <- load_files("LYMYLEUK", s$datafiles)
}
}

